fix - refetch on send new message and refactor
