+++ src/components/MessagesComponents/FormMessageSend.tsx
@@ -2,42 +2,19 @@ import { Send } from "lucide-react";
 import { Button } from "../ui/button";
 import { Input } from "../ui/input";
 import { useState } from "react";
-import { getToken } from "@/const/func";
-import axios from "axios";
 
 interface FormMessageSendProps {
-  receiver_id: number
+  onSendMessage: (message: string) => void;
 }
 
-function FormMessageSend({ receiver_id }: FormMessageSendProps) {
+function FormMessageSend({ onSendMessage }: FormMessageSendProps) {
   const [message, setMessage] = useState<string>("");
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-
-    // Ne rien faire si le message est vide
     if (!message.trim()) return;
-
-    try {
-      const token = getToken();
-      await axios.post(
-        `${import.meta.env.VITE_API_URL}/messages/send`,
-        {
-          content: message,
-          receiver_id: receiver_id,
-        },
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-
-      // Réinitialiser l'input après l'envoi
-      setMessage("");
-    } catch (error) {
-      console.error("Erreur lors de l'envoi du message :", error);
-    }
+    onSendMessage(message);
+    setMessage("");
   };
   return (
     <form className="flex items-center" onSubmit={handleSubmit}>
@@ -49,7 +26,7 @@ function FormMessageSend({ receiver_id }: FormMessageSendProps) {
       />
       <Button type="submit" disabled={!message.trim()}>
         Envoyer
-        <Send className="h-5 w-5 mr-1" />
+        <Send className="w-5 h-5 mr-1" />
       </Button>
     </form>
   );
