+++ src/components/MessagesComponents/MessagesList.tsx
@@ -1,10 +1,7 @@
 import { formatDistanceToNow, setDefaultOptions } from "date-fns";
 import { ScrollArea } from "../ui/scroll-area";
-import { useCallback, useEffect, useRef, useState } from "react";
-import { getToken } from "@/const/func";
-import axios from "axios";
+import { useEffect, useRef } from "react";
 import { fr } from "date-fns/locale";
-import { useWebSocket } from "../context/WebSocketContext";
 import { Check, CheckCheck } from "lucide-react";
 import { Conversation, Message } from "@/types/MessageTypes";
 
@@ -12,51 +9,14 @@ setDefaultOptions({ locale: fr });
 
 interface MessagesListProps {
   selectedConversation: Conversation | null;
+  messages: Message[];
 }
 
 function MessagesList({
   selectedConversation,
+  messages,
 }: MessagesListProps) {
-  const [messages, setMessages] = useState<Message[]>([]);
   const messageListRef = useRef<HTMLDivElement | null>(null); // Référence pour la ScrollArea
-  const { socket } = useWebSocket();
-
-  const fetchMessages = useCallback(async () => {
-    if (!selectedConversation) return;
-    try {
-      const token = getToken();
-      const response = await axios.get(
-        `${import.meta.env.VITE_API_URL}/messages/conversation/${
-          selectedConversation.user_id
-        }`,
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      setMessages(response.data);
-    } catch (err) {
-      console.error("Failed to fetch messages:", err);
-    }
-  }, [selectedConversation]);
-
-  useEffect(() => {
-    fetchMessages();
-  }, [fetchMessages]);
-
-  useEffect(() => {
-    if (!socket) return;
-
-    socket.on('newMessage', () => {
-      console.log('nouveau message dans MessagesList');
-      fetchMessages();
-    });
-
-    return () => {
-      socket.off('newMessage');
-    };
-  }, [socket, fetchMessages]);
 
   useEffect(() => {
     console.log("scroll to bottom");
